+++ apps/api/db/query/polling_unit_updates.sql
@@ -67,7 +67,7 @@ LEFT JOIN c_states s ON pu.state_id = s.id
 LEFT JOIN lgas l ON pu.lga_id = l.id
 WHERE 
   (sqlc.narg('election_group_id')::bigint IS NULL OR pu.election_group_id = sqlc.narg('election_group_id'))
-  AND (sqlc.narg('party_id')::bigint IS NULL OR pu.party_id = sqlc.narg('party_id'))
+  AND (sqlc.narg('party_id')::smallint IS NULL OR pu.party_id = sqlc.narg('party_id'))
   AND (sqlc.narg('polling_unit_id')::int IS NULL OR pu.polling_unit_id = sqlc.narg('polling_unit_id'))
   AND (sqlc.narg('user_id')::bigint IS NULL OR pu.user_id = sqlc.narg('user_id'))
   AND (sqlc.narg('state_id')::smallint IS NULL OR pu.state_id = sqlc.narg('state_id'))
